requirements.txt
# Core
TTS==0.22.0
torch==2.1.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.16.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
torchaudio==2.1.0+cpu -f https://download.pytorch.org/whl/torch_stable.html

# Video
moviepy==1.0.3

# YouTube
google-api-python-client
google-auth-oauthlib
google-auth-httplib2

# Utils
python-dotenv
PyYAML